@@ -13,7 +13,7 @@ class Poetry extends StatefulWidget {
 }
 
 class _PoetryState extends State<Poetry> with AutomaticKeepAliveClientMixin {
-  List list;
+  List ? list;
   int pageNum = 1;
 
   @override
@@ -30,12 +30,12 @@ class _PoetryState extends State<Poetry> with AutomaticKeepAliveClientMixin {
           ? Empty('暂无数据~')
           : ListView.separated(
               padding: EdgeInsets.all(16),
-              itemCount: this.list.length,
+              itemCount: this.list!.length,
               separatorBuilder: (BuildContext context, int index) => Container(
                     height: 12,
                   ),
               itemBuilder: (BuildContext context, int index) {
-                var item = this.list[index];
+                var item = this.list![index];
                 return PoetryItem(item);
               }),
     );
@@ -61,7 +61,7 @@ class _PoetryState extends State<Poetry> with AutomaticKeepAliveClientMixin {
       List list = data['list'];
       this.list = this.list ?? [];
       setState(() {
-        this.list.addAll(list);
+        this.list!.addAll(list);
       });
     }
   }
