@@ -27,15 +27,15 @@ class MyApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return BotToastInit(
-        child: MaterialApp(
-      title: '开眼视频',
+    return MaterialApp(
+      title: '听诗吧',
+      builder: BotToastInit(),
       theme: ThemeData(
           primarySwatch: Colors.red,
           primaryColor: Color(0xFFd46550),
           indicatorColor: Colors.white),
       navigatorObservers: [BotToastNavigatorObserver()],
       home: Index(),
-    ));
+    );
   }
 }
